% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fst.R
\name{fst}
\alias{fst}
\title{Fast S-Transform.}
\usage{
fst(xt, dt = 0.01, gauss = NA)
}
\arguments{
\item{xt}{Equally-sampled input time series. Must convert to numeric vector.}

\item{dt}{Sample interval, in seconds. Default is 0.01.}

\item{gauss}{TRUE for Gaussian windows, FALSE for Box windows. Default is TRUE.}
}
\value{
The complex S-transform of the data.
}
\description{
\code{fst} is used to compute the fast S-transform of a real,
univariate time series.
}
\details{
If Gaussian windows are selected, the results are discrete S-transforms.
If Box windows are selected, the results are discrete orthonormal S-transforms,
with critical sampling in both frequency and time/space dimensions.

This function is a wrapper for the Generalized Fourier Family Transform (GFT)
library \url{https://sourceforge.net/projects/fst-uofc/}.
}
\seealso{
\url{https://en.wikipedia.org/wiki/S_transform}
\url{http://ieeexplore.ieee.org/document/492555/}
\url{http://www.sciencedirect.com/science/article/pii/S1051200406000546}
\url{http://ieeexplore.ieee.org/document/5184926/}
\url{https://sourceforge.net/projects/fst-uofc/}
}
\keyword{ts}

