% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ifst.R
\name{ifst}
\alias{ifst}
\title{Fast Inverse S-Transform}
\usage{
ifst(dst, dt = 0.01, gauss = NA)
}
\arguments{
\item{dst}{Fast S-transform. Must convert to complex vector.}

\item{dt}{Sample interval, in seconds. Default is 0.01.}

\item{gauss}{TRUE for Gaussian windows, FALSE for Box windows. Default is TRUE.}
}
\value{
A list including the complex time series of the data, and
the window type used.
}
\description{
\code{ifst} is used to compute the fast inverse S-transform of a complex,
univariate S-transform.
}
\details{
If Gaussian windows are selected, the results are discrete S-transforms.
If Box windows are selected, the results are discrete orthonormal S-transforms,
with critical sampling in both frequency and time/space dimensions.

This function is a wrapper for the New Generalized Fourier Family Transform (NGFT)
library, a reimplementation of the Brown et al (2009) GFT library.
}
\seealso{
\itemize{
\item \href{https://en.wikipedia.org/wiki/S_transform}{Wikipedia}
\item \href{http://ieeexplore.ieee.org/document/492555/}{Stockwell et al, 1996}
Localization of the complex spectrum: the S transform.
\item \href{http://www.sciencedirect.com/science/article/pii/S1051200406000546}{Stockwell, 2007}
A basis for efficient representation of the S-transform.
\item \href{http://ieeexplore.ieee.org/document/5184926/}{Brown et al, 2009}
A General Description of Linear Time-Frequency Transforms
and Formulation of a Fast, Invertible Transform That Samples the Continuous
S-Transform Spectrum Nonredundantly.
\item \href{https://sourceforge.net/projects/fst-uofc/}{Original GFT library at SourceForge}
}
}
\keyword{ts}

